requirements.txt
coc.py >= 3.2.1
discord.py >= 2.3.0
pytest >= 7.3.2
pytest-asyncio >= 0.23.0
pyyaml >= 6.0
